requirements.txt
accelerate==1.6.0
altair==5.5.0
any-api==0.1.0.12
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
diffusers==0.33.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpc-gateway==0.0.1
grpcio==1.71.0
grpcio-tools==1.71.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pait==1.0.0a6
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==5.29.4
protobuf-to-pydantic==0.1.7.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==1.10.22
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.45.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
zipp==3.21.0
